feat: add entity type statistics bar chart to result overview
